scripts/requirements.txt
numpy
Pillow
ImageHash
opencv-python 